Other projects now use MFC as static library and now compile and work again. Implemented Game::indexToUnit in the server code and BWAPIClient.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2218 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BWAPI/GameImpl.cpp
@@ -1584,7 +1584,7 @@ namespace BWAPI
     return this->unitArray[i];
   }
   //----------------------------------------------------------------------------------------------------------
-  Unit *GameImpl::indexToUnit(short unitIndex)
+  Unit* GameImpl::indexToUnit(int unitIndex)
   {
     int i = (unitIndex & 0x7FF);
     if ( i < 1700 && this->unitArray[i]->canAccess() )
